Refactor product handler package
