+++ cmd/server/handlers/products/product.go
@@ -1,4 +1,4 @@
-package handlers
+package products
 
 import (
 	"fmt"
